Add a dropdown to select between workspaces.
There is still some implementation to do for this. The correct workspace needs to be passed to the chatbot. This requires both back and frontend changes. Additionally, the styling isn't ideal. The header could do with being more responsive.

Thought needs to go into how this works before the user is logged in
too.