Switch between workspaces.
Frontend changes include event handling for selecting workspaces, using
the introduction message for the chatbot and passing the namespace to
the backend via the chatbot api call.

Backend changes simply involve allowing the namespace as an argument.

Unfortunately the test for the selector isn't working. It's very tricky
to test select components in react. For whatever reason, I cannot
simulate the correct event to trigger the onValuechange at the top level
of the Select in WorkspaceList.tsx.

However, this is the only event that seems to actually fire when the
user selects something.

I'll remove the test in the next commit.